Cleans ts_trace API list body into tibble

Cleans ts_trace API list body into tibble

Usage

clean_trace_list(
  responsebody,
  data_type,
  gauge = NA,
  return_timezone = "UTC",
  .errorhandling = "stop"
)

Arguments

responsebody

response body from the API call to get_ts_traces

data_type

the data_type used to calculate the statistic over the interval, glued on for record

gauge

character gauge name- allows building an informative error-handled output

return_timezone

character in OlsonNames() or one of three special cases: 'db_default', 'char' or 'raw'. Default 'UTC'. If 'db_default', uses the API default. If 'char', returns a string in the format 'YYYY-MM-DDTHH:MM:SS+TZ', having all needed info (and matching Kiwis returns). If 'raw', returns the time column as-is from the API (a 14-digit string of numbers 'YYYMMDDHHMMSS')

"sxvlist" returns a list with an separate tibble for each site x variable combination

.errorhandling

as in foreach::foreach() (but handled in api_error_catch()) Default 'stop'. Made available here primarily to use 'pass' so big requests don't die due to API errors. Be careful- those errors then just get passed and so the data will be missing.

Value

a tibble with the rectangled response