Get the list of timeseries available

This is the API used by Australian Bureau of Meteorology and many others. For consistency with similar state functions using Kisters Hydstra, I have kept the station_no argument with the same name. Any of the return fields can be searched though, using extra_list. The equivalent state (Hydstra) function is get_variable_list() (to a close approximation). The available return fields (and thus factors that can be filtered) are 'station_name', 'station_no', 'station_id', 'ts_id', 'ts_name', 'ts_path', 'parametertype_id', 'parametertype_name', where the station_no argument matches station_no for consistency across state functions. Important- There is an additional returnfield option, coverage, which returns the period of record. It is included by default, but if you limit returnfields, you must include it to get the time period. There may be additional undocumented returnfield options; it appears that most of what is returned by getStationList is available, e.g. 'station_latitude'

Usage

getTimeseriesList(
  portal,
  station_no = NULL,
  returnfields = "default",
  extra_list = list(NULL),
  return_timezone = "UTC"
)

Arguments

portal

URL to Kisters KiWIS database.

station_no

gauge numbers, as site_list in the Hydstra functions. There are many other fields that can be used to filter and select records, but this seems to be most common and so we give it special treatment. For others, use see extra_list

returnfields

default 'default', otherwise 'all' to get everything available, or comma-separated string of fields to return. Important- 'coverage' is required to return start and end dates. It is included by default. Including 'ca_sta' will give the DATA_OWNER_NAME, which is often useful. Full list for each function available from Kisters docs

extra_list

a named list, see getStationList(), with a special note that here we can include a 'timezone' argument that determines the timezone the API returns in. This is dangerous, since the API ingests dates in its own default timezone and that is inferred from the return in the absence of the ability to extract it. Thus, including a timezone in extra_list may yield unexpected outcomes when requesting dates. A better option is to use return_timezone to adjust the return values. That said, it may be that some databases return gauge-local tzs, which won't be allowed to be concatenated. A solution would be to just work in UTC with timezone = 'UTC' in extralist to make all outputs on the same tz.

return_timezone

character in OlsonNames() or one of three special cases: 'db_default', 'char' or 'raw'. Default 'UTC'. If 'db_default', uses the API default. BOM defaults to +10. If 'char' or 'raw', returns the time column as-is from the API (A string in the format 'YYYY-MM-DDTHH:MM:SS+TZ')

Value

A tibble of information about available timeseries.Times are POSIXct in UTC by default.