Gets the variables available for sites and datasources

Gets the variables available for sites and datasources

Usage

get_variable_list(portal, site_list, datasource, return_timezone = "UTC")

Arguments

portal

character for the data portal (case insensitive). Default 'victoria'

site_list

character site code, either a single site code "sitenumber", comma-separated codes in a single string "sitenumber1, sitenumber2, or a vector of site codes c("sitenumber1", "sitenumber2")

datasource

as in get_ts_traces(), but can be a vector. As far as I can tell, options are 'A', 'TELEM', 'TELEMCOPY'. If multiple, c('A', 'TELEM')

return_timezone

character in OlsonNames() or one of three special cases: 'db_default', 'char' or 'raw'. Default 'UTC'. If 'db_default', uses the API default. If 'char', returns a string in the format 'YYYY-MM-DDTHH:MM:SS+TZ', having all needed info (and matching Kiwis returns). If 'raw', returns the time column as-is from the API (a 14-digit string of numbers 'YYYMMDDHHMMSS')

"sxvlist" returns a list with an separate tibble for each site x variable combination

Value

a tibble of the variables for each site and datasource.

Examples

v2 <- get_variable_list(portal = 'vic',
  site_list = "233217, 405328, 405331, 405837",
  datasource = c("A", "TELEM")
)